Changed local float vars to double in WeatherRock.
Changed
-------
* Changed float variables to double in function WeatherRock. This fixes
  a problem that resulted in the error message:
      "rock & behind both have no rocktype"
  When we switch the global floats to doubles, comparisons within this
  function nolonger worked.

+++ cem.c
@@ -2028,7 +2028,7 @@ WeatherRock (int j)
 /* Weathering rate changes along the shore, 'f' = fast weather, 's' = slow weathering			*/
 /* Function uses (but does not change) TypeOfRock[][] to determine rate					*/
 {
-  float WeatheringRatePerYear,
+  double WeatheringRatePerYear,
     AngleFactor,
     CurrentWeatherCoeff,
     mslope,
